fix: changed package name from com.rishav to com.dscvit
