+++ app/src/main/java/com/dscvit/gidget/common/Security.kt
@@ -1,0 +1,9 @@
+package com.dscvit.gidget.common
+
+import com.dscvit.gidget.BuildConfig
+
+class Security {
+    companion object {
+        fun getToken(): String = BuildConfig.TOKEN
+    }
+}
