--- app/src/main/java/com/rishav/gidget/common/Security.kt
@@ -1,9 +1,0 @@
-package com.rishav.gidget.common
-
-import com.rishav.gidget.BuildConfig
-
-class Security {
-    companion object {
-        fun getToken(): String = BuildConfig.TOKEN
-    }
-}
