--- app/src/main/java/com/rishav/gidget/realm/SignUp.kt
@@ -1,13 +1,0 @@
-package com.rishav.gidget.realm
-
-import io.realm.RealmObject
-import io.realm.annotations.PrimaryKey
-
-open class SignUp : RealmObject() {
-    @PrimaryKey
-    var email: String? = null
-
-    var name: String = ""
-    var photoUrl: String = ""
-    var username: String = ""
-}
