--- app/src/main/java/com/rishav/gidget/retrofit/RetroFitClient.kt
@@ -1,18 +1,0 @@
-package com.rishav.gidget.retrofit
-
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
-
-object RetroFitClient {
-    private var retrofit: Retrofit? = null
-
-    fun getClient(baseUrl: String): Retrofit {
-        if (retrofit == null) {
-            retrofit = Retrofit.Builder()
-                .baseUrl(baseUrl)
-                .addConverterFactory(MoshiConverterFactory.create())
-                .build()
-        }
-        return retrofit!!
-    }
-}
